deprecate Text & Badge
