+++ packages/wix-ui-core/src/components/Badge/Badge.e2e.ts
@@ -12,6 +12,6 @@ describe('Badge', () => {
     const driver = badgeTestkitFactory({dataHook});
 
     return waitForVisibilityOf(driver.element(), 'Cannot find Badge')
-      .then(() => expect(driver.getTextContent()).toBe('I\'m a Badge!'));
+      .then(() => expect(driver.text()).toBe('I\'m a Badge!'));
   });
 });
