+++ packages/wix-ui-core/src/components/Badge/Badge.protractor.driver.ts
@@ -1,6 +1,6 @@
 export const badgeDriverFactory = component => ({
-  /** returns the component element */
-  element: () => component,
-  /** returns the component text */
-  getTextContent: () => component.getText(),
-});
+    /** returns the component element */
+    element: () => component,
+    /** returns the component text */
+    text: () => component.getText()
+  });
