+++ packages/wix-ui-core/src/components/Badge/Badge.tsx
@@ -1,31 +1,22 @@
 import * as React from 'react';
 import {object, any} from 'prop-types';
-import {createHOC} from '../../createHOC';
-
-export type BadgeClasses = {
-  badge: string;
-};
+import style from './Badge.st.css';
 
 export interface BadgeProps {
-  classes?: BadgeClasses;
   children?: React.ReactNode;
 }
 
 /**
  * Badge
  */
-const Badge: React.SFC<BadgeProps> = ({classes, children}) => (
-  <span className={classes.badge}>
-    {children}
+export const Badge: React.SFC<BadgeProps> = props => (
+  <span {...style('root', {}, props)}>
+    {props.children}
   </span>
 );
 
 Badge.displayName = 'Badge';
 Badge.propTypes = {
-  /** Classes object */
-  classes: object.isRequired,
   /** Any node to be rendered (usually text node) */
   children: any
 };
-
-export default createHOC(Badge);
