+++ packages/wix-ui-core/src/components/Badge/index.ts
@@ -1,8 +1,1 @@
-import * as React from 'react';
-import BadgeComponent, {BadgeProps} from './Badge';
-import {styles} from './styles';
-import {withClasses, ThemedComponentProps} from 'wix-ui-jss';
-import {WixComponentProps} from '../../createHOC';
-
-export {BadgeProps};
-export const Badge = withClasses(BadgeComponent, styles) as React.ComponentClass<BadgeProps & ThemedComponentProps & WixComponentProps>;
+export {Badge, BadgeProps} from './Badge';
