--- packages/wix-ui-core/src/components/StylableBadge/Badge.driver.ts
@@ -1,12 +1,0 @@
-export const badgeDriverFactory = ({element}) => {
-    return {
-      /** checks if element exists */
-      exists: () => !!element,
-      /** returns elements type attribute */
-      getType: () => element.getAttribute('type'),
-      /** returns elements innerHtml */
-      getContent: () => element.innerHTML,
-      /** returns elements innerText */
-      text: () => element.textContent,
-    };
-  };
