--- packages/wix-ui-core/src/components/StylableBadge/Badge.protractor.driver.ts
@@ -1,6 +1,0 @@
-export const badgeDriverFactory = component => ({
-    /** returns the component element */
-    element: () => component,
-    /** returns the component text */
-    text: () => component.getText()
-  });
