--- packages/wix-ui-core/src/components/StylableBadge/Badge.spec.tsx
@@ -1,51 +1,0 @@
-import * as React from 'react';
-import {badgeDriverFactory} from './Badge.driver';
-import {isEnzymeTestkitExists} from 'wix-ui-test-utils/enzyme';
-import {createDriverFactory} from 'wix-ui-test-utils/driver-factory';
-import {isTestkitExists} from 'wix-ui-test-utils/vanilla';
-import {stylableBadgeTestkitFactory as badgeTestkitFactory} from '../../testkit';
-import {stylableBadgeTestkitFactory as enzymeBadgeTestkitFactory} from '../../testkit/enzyme';
-import {mount} from 'enzyme';
-import {Badge} from './';
-
-describe('Badge', () => {
-
-  const createDriver = createDriverFactory(badgeDriverFactory);
-
-  describe('children', () => {
-    it('should be rendered', () => {
-      const content = (
-        <div>
-          <span>Delete</span>
-          <i>?</i>
-        </div>
-      );
-      const driver = createDriver(<Badge>{content}</Badge>);
-      expect(driver.getContent()).toContain(mount(content).html());
-    });
-
-    it('should get the content text', () => {
-      const content = (
-        <div>
-          <span>Delete</span>
-          <i>?</i>
-        </div>
-      );
-      const driver = createDriver(<Badge>{content}</Badge>);
-
-      expect(driver.text()).toMatch('Delete?');
-    });
-  });
-
-  describe('testkit', () => {
-    it('should exist', () => {
-      expect(isTestkitExists(<Badge/>, badgeTestkitFactory)).toBe(true);
-    });
-  });
-
-  describe('enzyme testkit', () => {
-    it('should exist', () => {
-      expect(isEnzymeTestkitExists(<Badge/>, enzymeBadgeTestkitFactory, mount)).toBe(true);
-    });
-  });
-});
