--- packages/wix-ui-core/src/components/StylableBadge/index.ts
@@ -1,1 +1,0 @@
-export {Badge, BadgeProps} from './Badge';
