--- packages/wix-ui-core/src/components/StylableText/index.ts
@@ -1,1 +1,0 @@
-export {Text, TextProps} from './Text';
