+++ packages/wix-ui-core/src/components/Text/Text.driver.ts
@@ -1,21 +1,19 @@
-const getClasses = (element) => {
-  const rawCssClasses = element.getAttribute('class');
-  let cssClasses = [];
-  if (rawCssClasses !== null) {
-    cssClasses = element.getAttribute('class').split(' ');
-  }
-  return cssClasses;
-};
+import {StylableDOMUtil} from 'stylable/test-utils';
+import style from './Text.st.css';
 
 export const textDriverFactory = ({element}) => {
+  const stylableDOMUtil = new StylableDOMUtil(style);
+
   return {
     /** check if element exists */
     exists: () => !!element,
     /** check if component has ellipsis */
-    isEllipsis: () => getClasses(element).indexOf('ellipsis') !== -1,
+    isEllipsis: () => stylableDOMUtil.hasStyleState(element, 'ellipsis'),
     /** check if element has title attribute */
     hasTitleAttribute: () => element.getAttribute('title') !== null,
-    /** return text font-family value */
-    getFontFamily: () => window.getComputedStyle(element).fontFamily
+    /** get the rendered tag name */
+    getTagName: () => element.tagName,
+    /** get the rendered content */
+    getChildren: () => element.innerHTML
   };
 };
