+++ packages/wix-ui-core/src/components/Text/Text.protractor.driver.ts
