--- packages/wix-ui-core/src/components/Text/theme.ts
@@ -1,19 +1,0 @@
-export type TextTheme = {
-  fontFamily?: string;
-  fontSize?: string;
-  lineHeight?: string;
-  color?: string;
-  textTransform?: string;
-  letterSpacing?: string;
-  margin?: string;
-};
-
-export const core: TextTheme = {
-  fontFamily: 'Arial',
-  fontSize: '12px',
-  lineHeight: 'normal',
-  color: '#000',
-  textTransform: 'none',
-  letterSpacing: 'normal',
-  margin: '0'
-};
