+++ packages/wix-ui-core/stories/Text/index.ts
@@ -1,13 +1,13 @@
 import 'react';
 import createStory from '../create-story';
 
-import {Text} from '../../src/components/StylableText';
-import * as TextSource from '!raw-loader!../../src/components/StylableText/Text.tsx';
+import {Text} from '../../src/components/Text';
+import * as TextSource from '!raw-loader!../../src/components/Text/Text.tsx';
 
 export const story = () => createStory({
   category: 'Components',
   name: 'Text',
-  storyName: 'StylableText',
+  storyName: 'Text',
   component: Text,
   componentProps: (setState) => ({
     'data-hook': 'storybook-text',
